CHROMETWO/requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain-google-genai==0.0.6
browser-use==0.0.14
pydantic==2.5.2
python-dotenv==1.0.0
